+++ qa/rpc-tests/regtest_signrawtransaction.py
@@ -20,8 +20,7 @@ class RegtestSignrawtransactionTest (BitcoinTestFramework):
 
         # Create and sign Sapling transaction.
         # If the incorrect consensus branch id is selected, there will be a signing error. 
-        opid = self.nodes[1].z_sendmany(taddr,
-            [{'address': zaddr1, 'amount': 1}])
+        opid = self.nodes[1].z_sendmany(taddr, [{'address': zaddr1, 'amount': 1}], 1)
         wait_and_assert_operationid_status(self.nodes[1], opid)
 
 if __name__ == '__main__':