+++ src/wallet/asyncrpcoperation_sendmany.h
@@ -48,6 +48,7 @@ public:
         ZTXOSelector ztxoSelector,
         std::vector<SendManyRecipient> recipients,
         int minDepth,
+        unsigned int anchorDepth,
         TransactionStrategy strategy,
         CAmount fee = DEFAULT_FEE,
         UniValue contextInfo = NullUniValue);
@@ -72,6 +73,7 @@ private:
     ZTXOSelector ztxoSelector_;
     std::vector<SendManyRecipient> recipients_;
     int mindepth_{1};
+    unsigned int anchordepth_{nAnchorConfirmations};
     CAmount fee_;
     UniValue contextinfo_;     // optional data to include in return value from getStatus()