+++ src/wallet/wallet.h
@@ -53,7 +53,7 @@ extern unsigned int nTxConfirmTarget;
 extern bool bSpendZeroConfChange;
 extern bool fSendFreeTransactions;
 extern bool fPayAtLeastCustomFee;
-extern unsigned int nOrchardAnchorConfirmations;
+extern unsigned int nAnchorConfirmations;
 
 static const unsigned int DEFAULT_KEYPOOL_SIZE = 100;
 //! -paytxfee default
@@ -78,8 +78,10 @@ static const unsigned int WITNESS_CACHE_SIZE = MAX_REORG_LENGTH + 1;
 
 //! Amount of entropy used in generation of the mnemonic seed, in bytes.
 static const size_t WALLET_MNEMONIC_ENTROPY_LENGTH = 32;
-//! -orchardanchorconfirmations default
-static const unsigned int DEFAULT_ORCHARD_ANCHOR_CONFIRMATIONS = 1;
+//! -anchorconfirmations default
+static const unsigned int DEFAULT_ANCHOR_CONFIRMATIONS = 3;
+// Default minimum number of confirmations for note selection
+static const unsigned int DEFAULT_NOTE_CONFIRMATIONS = 10;
 
 extern const char * DEFAULT_WALLET_DAT;
 
@@ -1778,16 +1780,19 @@ public:
     bool IsSproutNullifierFromMe(const uint256& nullifier) const;
     bool IsSaplingNullifierFromMe(const uint256& nullifier) const;
 
-    void GetSproutNoteWitnesses(
+    bool GetSproutNoteWitnesses(
          const std::vector<JSOutPoint>& notes,
+         unsigned int confirmations,
          std::vector<std::optional<SproutWitness>>& witnesses,
-         uint256 &final_anchor);
-    void GetSaplingNoteWitnesses(
+         uint256 &final_anchor) const;
+    bool GetSaplingNoteWitnesses(
          const std::vector<SaplingOutPoint>& notes,
+         unsigned int confirmations,
          std::vector<std::optional<SaplingWitness>>& witnesses,
-         uint256 &final_anchor);
+         uint256 &final_anchor) const;
     std::vector<std::pair<libzcash::OrchardSpendingKey, orchard::SpendInfo>> GetOrchardSpendInfo(
-        const std::vector<OrchardNoteMetadata>& orchardNoteMetadata) const;
+        const std::vector<OrchardNoteMetadata>& orchardNoteMetadata,
+        uint256 anchor) const;
 
     isminetype IsMine(const CTxIn& txin) const;
     CAmount GetDebit(const CTxIn& txin, const isminefilter& filter) const;
@@ -1966,7 +1971,7 @@ public:
                           std::vector<SaplingNoteEntry>& saplingEntriesRet,
                           std::vector<OrchardNoteMetadata>& orchardNotesRet,
                           const std::optional<NoteFilter>& noteFilter,
-                          int minDepth=1,
+                          int minDepth,
                           int maxDepth=INT_MAX,
                           bool ignoreSpent=true,
                           bool requireSpendingKey=true,